requirements.txt
pytelegrambotapi==4.14.0
python-dotenv==1.0.0
google-cloud-vision==3.4.5
SQLAlchemy==2.0.23
aiohttp==3.8.6
yookassa==2.4.0
Pillow==10.1.0
requests==2.31.0